Move metadata table to it's own class.
