+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/metadata/MetadataBrowserMainView.form
@@ -477,14 +477,10 @@
 
                   <Layout class="org.netbeans.modules.form.compat2.layouts.support.JScrollPaneSupportLayout"/>
                   <SubComponents>
-                    <Component class="javax.swing.JTable" name="segmentJTable">
-                      <Properties>
-                        <Property name="model" type="javax.swing.table.TableModel" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
-                          <Connection code="new IrisMetadataTableModel()" type="code"/>
-                        </Property>
-                      </Properties>
+                    <Component class="cfa.vo.iris.visualizer.stil.MetadataJTable" name="metadataJTable1">
                       <BindingProperties>
-                        <BindingProperty name="selectionModel" source="starTableList" sourcePath="${selectionModel}" target="segmentJTable" targetPath="selectionModel" updateStrategy="0" immediately="false"/>
+                        <BindingProperty name="selectedStarTables" source="Form" sourcePath="${selectedTables}" target="metadataJTable1" targetPath="selectedStarTables" updateStrategy="0" immediately="false"/>
+                        <BindingProperty name="selectionModel" source="starTableList" sourcePath="${selectionModel}" target="metadataJTable1" targetPath="selectionModel" updateStrategy="0" immediately="false"/>
                       </BindingProperties>
                     </Component>
                   </SubComponents>
